refactor(coordinator): extract hotkey supervisor/bridge loops into coordinator/hotkey_loops.rs — behavior unchanged, 446 tests green
Pure move of 39 free functions (hotkey supervisor loops, bridge loops, shortcut
wiring, window-hotkey fallback) from coordinator.rs into a new sibling submodule.
Functions made pub(super) and reached from the parent via 'use hotkey_loops::*;'
(same convention as dictation/qa/resources). No logic, signature, or public-API
changes; coordinator.rs 6961→5807 lines.

cargo check: 0 errors, 87 warnings (unchanged)
cargo clippy --no-deps: 0 errors
cargo test --lib: 446 passed
tsc --noEmit / npm run build: green
